implement analytics db service
